fix scan_tokens, handle new line and whitespace, handle c line of string

+++ core/include/error.h
@@ -1,5 +1,5 @@
+#include <stdio.h>
 
 void error(int line, char* message);
-void report_error(int line, char* message, int* hadError);
+void report_error(size_t line, char* message, int* hadError);
 const char* format_error(int line, const char* message);
-
